// src/shared/enums/accouts.enum.ts
/**
 * Copyright (c) 2025 Webapp Software Solutions. All rights reserved.
 * This file belongs to Webapp Software Solutions and is proprietary and confidential.
 * Unauthorized copying of this file, via any medium is strictly prohibited.
 * Website: https://webappssoft.com
 */

export enum UserRoles {
  ROOT = 'ROOT',
  ROOT_STAFF = 'ROOT_STAFF',
  ADMIN = 'ADMIN',
  STAFF = 'STAFF',
  USER = 'USER',
  SURVEY_MASTER = 'SURVEY_MASTER',
}